# arm64, bpf: add internal-only MOV instruction to resolve per-CPU addrs